// src/app/shared/models/filter-params.model.ts
export class FilterParams {
  riserTag: Array<{
    value: string;
    checked: boolean;
  }>;
  subAssembly: Array<{
    value: string;
    checked: boolean;
  }>;
  component: Array<{
    value: string;
    checked: boolean;
  }>;
  recordType: Array<{
    value: string;
    checked: boolean;
  }>;
}